Utilizing modularization pattern for Network
