+++ .idea/compiler.xml
@@ -1,6 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="CompilerConfiguration">
-    <bytecodeTargetLevel target="18" />
+    <bytecodeTargetLevel target="16">
+      <module name="TrainingApp.app" target="18" />
+    </bytecodeTargetLevel>
   </component>
 </project>
