+++ .idea/gradle.xml
@@ -12,6 +12,9 @@
           <set>
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
+            <option value="$PROJECT_DIR$/module" />
+            <option value="$PROJECT_DIR$/module/core" />
+            <option value="$PROJECT_DIR$/module/core/network" />
           </set>
         </option>
       </GradleProjectSettings>
