+++ app/build.gradle
@@ -78,6 +78,9 @@ dependencies {
     // for observeAsState
     implementation "androidx.compose.runtime:runtime-livedata:1.2.1"
 
+    // local Modules
+    implementation project(':module:core:network')
+
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation "androidx.compose.ui:ui:$compose_version"
     implementation 'androidx.compose.material3:material3:1.0.0-alpha01'
