+++ app/src/main/java/com/neudesic/myapplication/domain/network/DadJokeAPIService.kt
@@ -1,5 +1,6 @@
 package com.neudesic.myapplication.domain.network
 
+import com.neudesic.module.core.network.NetworkResult
 import com.neudesic.myapplication.data.repository.dto.DadJokeDTO
 import retrofit2.http.GET
 import retrofit2.http.Headers
