+++ build.gradle
@@ -17,6 +17,7 @@ plugins {
     id 'com.android.library' version '7.2.2' apply false
     id 'org.jetbrains.kotlin.android' version '1.7.0' apply false
     id 'com.google.dagger.hilt.android' version '2.43.2' apply false
+    id 'org.jetbrains.kotlin.jvm' version "$kotlin_version" apply false
 }
 
 task clean(type: Delete) {
