+++ module/core/network/src/main/java/com/neudesic/module/core/network/CoreNetworkModule.kt
@@ -1,0 +1,7 @@
+package com.neudesic.module.core.network
+
+import retrofit2.CallAdapter
+
+interface CoreNetworkModule {
+    fun networkResultCallAdapterFactory(): CallAdapter.Factory
+}
