+++ module/core/network/src/main/java/com/neudesic/module/core/network/CoreNetworkModuleImpl.kt
@@ -1,0 +1,7 @@
+package com.neudesic.module.core.network
+import com.neudesic.module.core.network.internal.NetworkResultCallAdapterFactory
+import retrofit2.CallAdapter
+
+class CoreNetworkModuleImpl():CoreNetworkModule {
+    override fun networkResultCallAdapterFactory(): CallAdapter.Factory = NetworkResultCallAdapterFactory.create()
+}
