+++ module/core/network/src/main/java/com/neudesic/module/core/network/internal/NetworkResultCallAdapter.kt
@@ -1,0 +1,17 @@
+package com.neudesic.module.core.network.internal
+
+import com.neudesic.module.core.network.NetworkResult
+import retrofit2.Call
+import retrofit2.CallAdapter
+import java.lang.reflect.Type
+
+internal class NetworkResultCallAdapter(
+    private val resultType: Type
+) : CallAdapter<Type, Call<NetworkResult<Type>>> {
+
+    override fun responseType(): Type = resultType
+
+    override fun adapt(call: Call<Type>): Call<NetworkResult<Type>> {
+        return NetworkResultCall(call)
+    }
+}
