+++ module/core/network/src/main/java/com/neudesic/module/core/network/internal/NetworkResultCallAdapterFactory.kt
@@ -1,0 +1,33 @@
+package com.neudesic.module.core.network.internal
+
+import com.neudesic.module.core.network.NetworkResult
+import retrofit2.Call
+import retrofit2.CallAdapter
+import retrofit2.Retrofit
+import java.lang.reflect.ParameterizedType
+import java.lang.reflect.Type
+
+internal class NetworkResultCallAdapterFactory private constructor() : CallAdapter.Factory() {
+
+    override fun get(
+        returnType: Type,
+        annotations: Array<out Annotation>,
+        retrofit: Retrofit
+    ): CallAdapter<*, *>? {
+        if (getRawType(returnType) != Call::class.java) {
+            return null
+        }
+
+        val callType = getParameterUpperBound(0, returnType as ParameterizedType)
+        if (getRawType(callType) != NetworkResult::class.java) {
+            return null
+        }
+
+        val resultType = getParameterUpperBound(0, callType as ParameterizedType)
+        return NetworkResultCallAdapter(resultType)
+    }
+
+    companion object {
+        fun create(): NetworkResultCallAdapterFactory = NetworkResultCallAdapterFactory()
+    }
+}
