+++ settings.gradle
@@ -14,3 +14,4 @@ dependencyResolutionManagement {
 }
 rootProject.name = "TrainingApp"
 include ':app'
+include ':module:core:network'
